feat: PolarDB-PG 15 open source
This commit open sources PolarDB-PG 15 after PostgreSQL 15.8. The main
features include:

- basic CI supporting checks
- VFS and compute-storage seperated architecture
- buffer manager for primary-replica
- sync/async DDL for primary-replica
- logindex
- resource manager
- ring buffer and I/O optimizations
- backup/restore under shared-storage
- several popular extensions

+++ src/bin/pg_dump/pg_dumpall.c
@@ -1198,6 +1198,7 @@ dropDBs(PGconn *conn)
 		 */
 		if (strcmp(dbname, "template1") != 0 &&
 			strcmp(dbname, "template0") != 0 &&
+			strcmp(dbname, "polardb_admin") != 0 &&
 			strcmp(dbname, "postgres") != 0)
 		{
 			fprintf(OPF, "DROP DATABASE %s%s;\n",
@@ -1359,7 +1360,8 @@ dumpDatabases(PGconn *conn)
 		 * otherwise we'll merely restore their contents.  Other databases
 		 * should simply be created.
 		 */
-		if (strcmp(dbname, "template1") == 0 || strcmp(dbname, "postgres") == 0)
+		if (strcmp(dbname, "template1") == 0 || strcmp(dbname, "postgres") == 0 ||
+			strcmp(dbname, "polardb_admin") == 0)
 		{
 			if (output_clean)
 				create_opts = "--clean --create";
